add chart card and legend components for improved asset visualization
