--- asset_tracker_app/lib/widgets/profile_view/error_widget.dart
@@ -1,54 +1,0 @@
-import 'package:flutter/material.dart';
-
-/// A custom widget to display error messages
-///
-/// Provides a standardized error UI across the application
-class CustomErrorWidget extends StatelessWidget {
-  final String message;
-  final VoidCallback? onRetry;
-
-  /// Creates a custom error widget
-  ///
-  /// [message] is the error message to display
-  /// [onRetry] optional callback for retry functionality
-  const CustomErrorWidget({
-    super.key,
-    required this.message,
-    this.onRetry,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            const Icon(
-              Icons.error_outline,
-              size: 60,
-              color: Colors.red,
-            ),
-            const SizedBox(height: 16),
-            Text(
-              message,
-              textAlign: TextAlign.center,
-              style: const TextStyle(
-                fontSize: 16,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-            if (onRetry != null) ...[
-              const SizedBox(height: 16),
-              ElevatedButton(
-                onPressed: onRetry,
-                child: const Text('Tekrar Dene'),
-              ),
-            ],
-          ],
-        ),
-      ),
-    );
-  }
-}
